01-content-factory/python/requirements.txt
# Content Factory Agent - Python Dependencies

# LangGraph & LangChain
langgraph>=1.0.0
langchain>=0.3.0
langchain-anthropic>=0.3.0
langchain-google-genai>=2.0.0

# AI Providers
anthropic>=0.75.0
google-generativeai>=0.8.0

# Web Framework
fastapi>=0.120.0
uvicorn>=0.30.0
pydantic>=2.10.0

# HTTP Client
httpx>=0.28.0
aiohttp>=3.10.0

# Utils
python-dotenv>=1.0.0
